Fixed 404 issue and calling 3rd party API

+++ src/main/java/com/scaler/ProductService/services/ProductService.java
@@ -4,5 +4,5 @@ import com.scaler.ProductService.models.Product;
 
 public interface ProductService {
 
-    public Product getProductById(int id);
+    public Product getProductById(long id);
 }
